DSPCore: Rename namespace to match moved x64 JIT files
Maintains consistency between location and code
