+++ Source/Core/Core/DSP/DSPTables.h
@@ -69,7 +69,7 @@ struct param2_t
 struct DSPOPCTemplate
 {
   using InterpreterFunction = void (*)(UDSPInstruction);
-  using JITFunction = void (DSP::JIT::x86::DSPEmitter::*)(UDSPInstruction);
+  using JITFunction = void (JIT::x64::DSPEmitter::*)(UDSPInstruction);
 
   const char* name;
   u16 opcode;
