+++ Source/Core/Core/DSP/Jit/x64/DSPJitMisc.cpp
@@ -14,7 +14,7 @@ namespace DSP
 {
 namespace JIT
 {
-namespace x86
+namespace x64
 {
 // MRR $D, $S
 // 0001 11dd ddds ssss
@@ -195,6 +195,6 @@ void DSPEmitter::srbith(const UDSPInstruction opc)
   }
 }
 
-}  // namespace x86
+}  // namespace x64
 }  // namespace JIT
 }  // namespace DSP
